ST#730 - TK#4750 - Write Off upload changes

+++ PFSJava/src/main/java/com/pennant/backend/service/finance/impl/FinanceWriteoffServiceImpl.java
@@ -83,6 +83,7 @@ import com.pennant.backend.util.FinanceConstants;
 import com.pennant.backend.util.PennantApplicationUtil;
 import com.pennant.backend.util.PennantConstants;
 import com.pennant.backend.util.PennantJavaUtil;
+import com.pennant.backend.util.UploadConstants;
 import com.pennant.cache.util.AccountingConfigCache;
 import com.pennant.pff.accounting.model.PostingDTO;
 import com.pennant.pff.core.engine.accounting.AccountingEngine;
@@ -221,6 +222,9 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 				financeWriteoff.setProvisionedAmount(provision.getProvisionedAmt());
 			}
 
+			financeWriteoff.setFinID(finID);
+			financeWriteoff.setFinReference(finReference);
+			financeWriteoff.setWriteoffDate(SysParamUtil.getAppDate());
 			writeoffHeader.setFinanceWriteoff(financeWriteoff);
 		}
 
@@ -616,27 +620,7 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 		// Update Profit Details
 		// getProfitDetailsDAO().update(profitDetail, false);
 
-		// Schedule Details delete
-		// =======================================
-		listDeletion(finID, "_Temp");
-
-		// Fee charges deletion
-		List<AuditDetail> tempAuditDetailList = new ArrayList<AuditDetail>();
-		finFeeChargesDAO.deleteChargesBatch(finID, fd.getModuleDefiner(), false, "_Temp");
-
-		// Checklist Details delete
-		// =======================================
-		tempAuditDetailList.addAll(checkListDetailService.delete(fd, "_Temp", tranType));
-
-		// Finance Writeoff Details
-		financeWriteoffDAO.delete(finID, "_Temp");
-
 		String[] fields = PennantJavaUtil.getFieldDetails(new FinanceMain(), fm.getExcludeFields());
-		financeMainDAO.delete(fm, TableType.TEMP_TAB, false, true);
-
-		if (CollectionUtils.isNotEmpty(schdData.getFinServiceInstructions())) {
-			finServiceInstructionDAO.saveList(schdData.getFinServiceInstructions(), "");
-		}
 
 		// Extended field Render Details.
 		List<AuditDetail> details = fd.getAuditDetailMap().get("ExtendedFieldDetails");
@@ -646,30 +630,52 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 			auditDetails.addAll(details);
 		}
 
-		finServiceInstructionDAO.deleteList(finID, fd.getModuleDefiner(), "_Temp");
+		if (!StringUtils.equals(UploadConstants.FINSOURCE_ID_UPLOAD, header.getFinSource())) {
+			// Schedule Details delete
+			// =======================================
+			listDeletion(finID, "_Temp");
 
-		if (details != null && details.size() > 0) {
-			extendedFieldDetailsService.delete(header.getFinanceDetail().getExtendedFieldHeader(),
-					header.getFinanceDetail().getExtendedFieldRender().getReference(),
-					header.getFinanceDetail().getExtendedFieldRender().getSeqNo(), "_Temp",
-					auditHeader.getAuditTranType(), details);
+			// Fee charges deletion
+			List<AuditDetail> tempAuditDetailList = new ArrayList<AuditDetail>();
+			finFeeChargesDAO.deleteChargesBatch(finID, fd.getModuleDefiner(), false, "_Temp");
+
+			// Checklist Details delete
+			// =======================================
+			tempAuditDetailList.addAll(checkListDetailService.delete(fd, "_Temp", tranType));
+
+			// Finance Writeoff Details
+			financeWriteoffDAO.delete(finID, "_Temp");
+
+			financeMainDAO.delete(fm, TableType.TEMP_TAB, false, true);
+
+			finServiceInstructionDAO.deleteList(finID, fd.getModuleDefiner(), "_Temp");
+
+			if (details != null && details.size() > 0) {
+				extendedFieldDetailsService.delete(header.getFinanceDetail().getExtendedFieldHeader(),
+						header.getFinanceDetail().getExtendedFieldRender().getReference(),
+						header.getFinanceDetail().getExtendedFieldRender().getSeqNo(), "_Temp",
+						auditHeader.getAuditTranType(), details);
+			}
+
+			for (int i = 0; i < auditDetails.size(); i++) {
+				auditHeader.setErrorList(auditDetails.get(i).getErrorDetails());
+			}
+
+			auditHeader.setAuditDetail(
+					new AuditDetail(aAuditHeader.getAuditTranType(), 1, fields[0], fields[1], fm.getBefImage(), fm));
+			auditHeader.setAuditDetails(tempAuditDetailList);
+			auditHeader.setAuditModule("FinanceDetail");
+			auditHeaderDAO.addAudit(auditHeader);
+
 		}
 
-		for (int i = 0; i < auditDetails.size(); i++) {
-			auditHeader.setErrorList(auditDetails.get(i).getErrorDetails());
+		if (CollectionUtils.isNotEmpty(schdData.getFinServiceInstructions())) {
+			finServiceInstructionDAO.saveList(schdData.getFinServiceInstructions(), "");
 		}
-
-		// Adding audit as deleted from TEMP table
-		String auditTranType = aAuditHeader.getAuditTranType();
-
-		auditHeader.setAuditDetail(new AuditDetail(auditTranType, 1, fields[0], fields[1], fm.getBefImage(), fm));
-		auditHeader.setAuditDetails(tempAuditDetailList);
-		auditHeader.setAuditModule("FinanceDetail");
-		auditHeaderDAO.addAudit(auditHeader);
 
 		auditHeader.setAuditTranType(tranType);
-		auditTranType = auditHeader.getAuditTranType();
-		auditHeader.setAuditDetail(new AuditDetail(auditTranType, 1, fields[0], fields[1], fm.getBefImage(), fm));
+		auditHeader.setAuditDetail(
+				new AuditDetail(aAuditHeader.getAuditTranType(), 1, fields[0], fields[1], fm.getBefImage(), fm));
 		auditHeader.setAuditDetails(auditDetails);
 		auditHeader.setAuditModule("FinanceDetail");
 		auditHeaderDAO.addAudit(auditHeader);
@@ -677,11 +683,6 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 		auditHeader.getAuditDetail().setModelData(header);
 
 		finStageAccountingLogDAO.update(finID, fd.getModuleDefiner(), false);
-
-		// updating NPA Loans
-		if (FinanceConstants.CLOSE_STATUS_WRITEOFF.equals(fm.getClosingStatus())) {
-			assetClassificationService.doCloseLoan(finID);
-		}
 
 		logger.debug(Literal.LEAVING);
 		return auditHeader;
@@ -875,84 +876,11 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 
 		long finID = fm.getFinID();
 
-		FinanceMain tempFinanceMain = null;
-
-		if (fm.isWorkflow()) {
-			tempFinanceMain = financeMainDAO.getFinanceMainById(finID, "_Temp", false);
-		}
-
 		FinanceMain befFinanceMain = financeMainDAO.getFinanceMainById(finID, "", false);
-		FinanceMain oldFinanceMain = fm.getBefImage();
 
 		String[] errParm = new String[1];
 		String[] valueParm = new String[1];
 		valueParm[0] = fm.getFinReference();
-		errParm[0] = PennantJavaUtil.getLabel("label_FinReference") + ":" + valueParm[0];
-
-		if (fm.isNewRecord()) { // for New record or new record into work flow
-
-			if (!fm.isWorkflow()) {// With out Work flow only new
-				// records
-				if (befFinanceMain != null) { // Record Already Exists in the
-					// table then error
-					auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-							new ErrorDetail(PennantConstants.KEY_FIELD, "41001", errParm, valueParm), usrLanguage));
-				}
-			} else { // with work flow
-				if (fm.getRecordType().equals(PennantConstants.RECORD_TYPE_NEW)) { // if
-					// records type is new
-					if (befFinanceMain != null || tempFinanceMain != null) { // if
-						// records already exists in the main table
-						auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-								new ErrorDetail(PennantConstants.KEY_FIELD, "41001", errParm, valueParm), usrLanguage));
-					}
-				} else { // if records not exists in the Main flow table
-					if (befFinanceMain == null || tempFinanceMain != null) {
-						auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-								new ErrorDetail(PennantConstants.KEY_FIELD, "41005", errParm, valueParm), usrLanguage));
-					}
-				}
-			}
-		} else {
-			// for work flow process records or (Record to update or Delete with
-			// out work flow)
-			if (!fm.isWorkflow()) { // With out Work flow for update
-				// and delete
-
-				if (befFinanceMain == null) { // if records not exists in the
-					// main table
-					auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-							new ErrorDetail(PennantConstants.KEY_FIELD, "41002", errParm, valueParm), usrLanguage));
-				} else {
-					if (oldFinanceMain != null
-							&& !oldFinanceMain.getLastMntOn().equals(befFinanceMain.getLastMntOn())) {
-						if (StringUtils.trimToEmpty(auditDetail.getAuditTranType())
-								.equalsIgnoreCase(PennantConstants.TRAN_DEL)) {
-							auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-									new ErrorDetail(PennantConstants.KEY_FIELD, "41003", errParm, valueParm),
-									usrLanguage));
-						} else {
-							auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-									new ErrorDetail(PennantConstants.KEY_FIELD, "41004", errParm, valueParm),
-									usrLanguage));
-						}
-					}
-				}
-			} else {
-
-				if (tempFinanceMain == null) { // if records not exists in the
-					// Work flow table
-					auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-							new ErrorDetail(PennantConstants.KEY_FIELD, "41005", errParm, valueParm), usrLanguage));
-				}
-
-				if (tempFinanceMain != null && oldFinanceMain != null
-						&& !oldFinanceMain.getLastMntOn().equals(tempFinanceMain.getLastMntOn())) {
-					auditDetail.setErrorDetail(ErrorUtil.getErrorDetail(
-							new ErrorDetail(PennantConstants.KEY_FIELD, "41005", errParm, valueParm), usrLanguage));
-				}
-			}
-		}
 
 		// Checking , if Customer is in EOD process or not. if Yes, not allowed to do an action
 		int eodProgressCount = customerQueuingDAO.getProgressCountByCust(fm.getCustID());
@@ -1051,6 +979,11 @@ public class FinanceWriteoffServiceImpl extends GenericFinanceDetailService impl
 	@Override
 	public List<ManualAdvise> getPayableAdvises(long finID, String type) {
 		return this.manualAdviseDAO.getPaybleAdvises(finID, type);
+	}
+
+	@Override
+	public boolean isWriteoffLoan(long finID, String type) {
+		return this.financeWriteoffDAO.isWriteoffLoan(finID, type);
 	}
 
 	@Autowired
